Reorganise files into directories
