+++ src/main/java/duke/exception/DukeException.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.exception;
 
 /**
  * A specified Exception for this project.
