+++ src/test/java/duke/ToDoTest.java
@@ -4,6 +4,9 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 
+import duke.exception.DukeException;
+import duke.task.ToDo;
+
 public class ToDoTest {
 
     @Test
